feat(#S11P12D109-136): 리뷰 리팩터링 완료

+++ Colorpl-Backend/src/main/java/com/colorpl/review/repository/ReviewRepository.java
@@ -3,6 +3,9 @@ package com.colorpl.review.repository;
 import com.colorpl.review.domain.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReviewRepository extends JpaRepository<Review, Integer> {
+public interface ReviewRepository extends JpaRepository<Review, Long> {
+    // 사용자 이름으로 리뷰 찾기
+//    List<Review> findByUserId(Long reviewId);
 
+//    List<Review> findByScheduleMemberId(Integer memberId);
 }
